Server/Python/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.11
filelock==3.16.1
fsspec==2025.3.0
h11==0.14.0
huggingface-hub==0.29.3
idna==3.10
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
packaging==24.2
pillow==10.4.0
pydantic==2.10.6
pydantic_core==2.27.2
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
safetensors==0.5.3
sniffio==1.3.1
starlette==0.44.0
sympy==1.13.3
timm==1.0.15
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.33.0
